fix(graph): keep staff directory port manifest-safe
