+++ packages/auth/src/ports.ts
@@ -1,1 +1,7 @@
 export { customerBusinessAccountOnboardingRuntimePort } from "./customer-business-onboarding-runtime-port.js"
+export {
+  type ActiveStaffIdentity,
+  type StaffDirectoryLookupContext,
+  type StaffDirectoryRuntimeProvider,
+  staffDirectoryRuntimePort,
+} from "./staff-directory-port.js"
