+++ packages/auth/src/staff-directory-runtime-port.ts
@@ -1,41 +1,12 @@
-import { definePort } from "@voyant-travel/core/project"
 import { authUser } from "@voyant-travel/db/schema/iam"
-import type { VoyantDb } from "@voyant-travel/hono"
+import type {
+  ActiveStaffIdentity,
+  StaffDirectoryLookupContext,
+  StaffDirectoryRuntimeProvider,
+} from "./staff-directory-port.js"
 import type { TeamManagementRuntimeProvider } from "./team-management-runtime-port.js"
 
-export interface ActiveStaffIdentity {
-  userId: string
-  displayName: string
-}
-
-export interface StaffDirectoryLookupContext {
-  db: VoyantDb
-  bindings: Record<string, unknown>
-  requesterUserId: string
-}
-
-/** Auth-owned stable user identities exposed without copying profiles into consumers. */
-export interface StaffDirectoryRuntimeProvider {
-  isActiveStaff(context: StaffDirectoryLookupContext, userId: string): Promise<boolean>
-  listActiveStaff(
-    context: StaffDirectoryLookupContext,
-    input?: { userIds?: readonly string[] },
-  ): Promise<readonly ActiveStaffIdentity[]>
-}
-
-export const staffDirectoryRuntimePort = definePort<StaffDirectoryRuntimeProvider>({
-  id: "auth.staff-directory",
-  test(provider) {
-    if (
-      !provider ||
-      typeof provider !== "object" ||
-      typeof provider.isActiveStaff !== "function" ||
-      typeof provider.listActiveStaff !== "function"
-    ) {
-      throw new Error("auth.staff-directory provider must implement active staff lookups.")
-    }
-  },
-})
+export * from "./staff-directory-port.js"
 
 export function createTeamBackedStaffDirectory(
   team: TeamManagementRuntimeProvider,
